[add]タスク追加と削除をコンポーネント化
